fix: u8 for u16 in proving system id

+++ contracts/src/core/AlignedProofAggregationService.sol
@@ -111,14 +111,14 @@ contract AlignedProofAggregationService is
     /// - The function returns `true` if this Merkle root is known to correspond to a valid aggregated proof.
     ///
     /// @param merklePath The Merkle proof (sibling hashes) needed to reconstruct the Merkle root.
-    /// @param provingSystemId The id of the proving system (0 for SP1, 1 for RISC0).
+    /// @param provingSystemId The id of the proving system (1 for SP1, 2 for RISC0).
     /// @param programId The identifier for the ZK program (image_id in RISC0 or vk hash in SP1).
     /// @param publicInputs The public inputs bytes of the proof.
     ///
     /// @return bool Returns true if the computed Merkle root is a recognized valid aggregated proof.
     function verifyProofInclusion(
         bytes32[] calldata merklePath,
-        uint8 provingSystemId,
+        uint16 provingSystemId,
         bytes32 programId,
         bytes calldata publicInputs
     ) public view returns (bool) {
